backend/pyproject.toml
[tool.poetry]
name = "project-poetry"
version = "0.1.0"
description = ""
authors = ["houston03 <super.avel-2014@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiohappyeyeballs = "^2.4.4"
aiohttp = "^3.11.11"
aiohttp-retry = "^2.8.3"
aiosignal = "^1.3.2"
alembic = "^1.14.0"
amqp = "^5.3.1"
annotated-types = "^0.7.0"
anyio = "^4.7.0"
asyncpg = "^0.30.0"
attrs = "^24.3.0"
bcrypt = "^4.2.1"
billiard = "^4.2.1"
build = "^1.2.2.post1"
CacheControl = "^0.14.1"
celery = "^5.4.0"
certifi = "^2024.12.14"
cffi = "^1.17.1"
charset-normalizer = "^3.4.1"
cleo = "^2.1.0"
click = "^8.1.8"
click-didyoumean = "^0.3.1"
click-plugins = "^1.1.1"
click-repl = "^0.3.0"
colorama = "^0.4.6"
crashtest = "^0.4.1"
distlib = "^0.3.9"
dnspython = "^2.7.0"
dulwich = "^0.21.7"
email_validator = "^2.2.0"
fastapi = "^0.115.6"
fastjsonschema = "^2.21.1"
filelock = "^3.16.1"
frozenlist = "^1.5.0"
h11 = "^0.14.0"
httpcore = "^1.0.7"
httptools = "^0.6.4"
httpx = "^0.28.1"
idna = "^3.10"
installer = "^0.7.0"

jose = "^1.0.0"
keyring = "^24.3.1"
kombu = "^5.4.2"
Mako = "^1.3.8"
MarkupSafe = "^3.0.2"
more-itertools = "^10.5.0"
msgpack = "^1.1.0"
multidict = "^6.1.0"
packaging = "^24.2"
passlib = "^1.7.4"
pexpect = "^4.9.0"
pika = "^1.3.2"
pkginfo = "^1.12.0"
platformdirs = "^4.3.6"
poetry = "^1.8.5"
poetry-core = "^1.9.1"
poetry-plugin-export = "^1.8.0"
prompt_toolkit = "^3.0.48"
propcache = "^0.2.1"
ptyprocess = "^0.7.0"
pycparser = "^2.22"
pydantic = "^2.10.4"
pydantic_core = "^2.27.2"
PyJWT = "^2.10.1"
pyproject_hooks = "^1.2.0"
python-dateutil = "^2.9.0.post0"
python-dotenv = "^1.0.1"
pywin32-ctypes = "^0.2.3"
PyYAML = "^6.0.2"
rabbitmq = "^0.2.0"
RapidFuzz = "^3.11.0"
redis = "^5.2.1"
requests = "^2.32.3"
requests-toolbelt = "^1.0.0"
shellingham = "^1.5.4"
six = "^1.17.0"
sniffio = "^1.3.1"
SQLAlchemy = "^2.0.36"
starlette = "^0.41.3"
tomlkit = "^0.13.2"
trove-classifiers = "^2024.10.21.16"
typing_extensions = "^4.12.2"
tzdata = "^2024.2"
urllib3 = "^2.3.0"
uvicorn = "^0.34.0"
vine = "^5.1.0"
virtualenv = "^20.28.0"
watchfiles = "^1.0.3"
wcwidth = "^0.2.13"
websockets = "^14.1"
yarl = "^1.18.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
